extras: move fake GPS velocity/course helpers to frame_tf
Move calculate_velocity_ned() and course_over_ground_cdeg() into the
generic mavros::ftf helpers and drop the fake_gps_utils.hpp shim and its
dedicated test target. The fake GPS velocity and course tests now live in
the existing libmavros-frame-conversions test.
